使用品質提升
刪除 登出 註冊 防呆

+++ scripts/viewproduct.js
@@ -39,12 +39,15 @@ onAuthStateChanged(auth, async (user) => {
         btn.forEach((b) => {
             b.addEventListener('click', (e) => {
                 e.preventDefault();
-                var docRef = doc(db, 'Product', b.id);
-                deleteDoc(docRef)
-                .then(() => {
-                    alert("已成功刪除!");
-                    location.reload();
-                });
+                let sure = confirm("是否確定刪除？");
+                if (sure) {
+                    var docRef = doc(db, 'Product', b.id);
+                    deleteDoc(docRef)
+                    .then(() => {
+                        alert("已成功刪除!");
+                        location.reload();
+                    });
+                }
             });
         });
